+++ libs/auth.py
@@ -3,8 +3,8 @@ from hashlib import sha224
 from tornado import gen
 from psycopg2 import extras
 
-from basehandler import BaseHandler
-from service_functions import showError, default_headers, log
+from libs.basehandler import BaseHandler
+from libs.service_functions import showError, default_headers, log
 
 class Auth(BaseHandler):
 	"""
