+++ libs/fapi.py
@@ -3,8 +3,8 @@ from json import loads, dumps
 from psycopg2 import extras
 from tornado import gen
 
-from basehandler import BaseHandler
-from service_functions import showError, default_headers, log
+from libs.basehandler import BaseHandler
+from libs.service_functions import showError, default_headers, log
 from settings import maindomain, primaryAuthorization
 
 def savefile(self):
